# docs/openapi-specification.yaml
openapi: 3.0.3
info:
  title: Thai Auto Insurance API
  version: 1.0.0
  description: |
    Comprehensive REST API for Thai Auto Insurance application supporting policy management, claims processing, and customer service operations.
    
    ## Features
    - Multi-language support (Thai/English)
    - JWT-based authentication
    - Thai-specific validations (National ID, phone numbers, addresses)
    - Real-time payment processing
    - Comprehensive claims management
    - Regulatory compliance (OIC requirements)
    
  contact:
    name: API Support Team
    email: api-support@autoinsurance.co.th
    url: https://docs.autoinsurance.co.th
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: https://autoinsurance.co.th/terms
servers:
  - url: https://api.autoinsurance.co.th/api/v1
    description: Production server
  - url: https://api-staging.autoinsurance.co.th/api/v1
    description: Staging server
  - url: https://api-dev.autoinsurance.co.th/api/v1
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and return JWT tokens
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              customer_login:
                summary: Customer login
                value:
                  email: "somchai@example.com"
                  password: "SecurePass123!"
                  deviceInfo:
                    deviceId: "uuid-device-id"
                    deviceType: "mobile"
                    appVersion: "1.0.0"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Create new customer account
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh JWT token
      description: Get new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request password reset
      description: Send password reset email
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset password
      description: Reset password using reset token
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # Customer Management APIs
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "somchai@example.com"
        password:
          type: string
          minLength: 8
          example: "SecurePass123!"
        deviceInfo:
          $ref: '#/components/schemas/DeviceInfo'
    DeviceInfo:
      type: object
      properties:
        deviceId:
          type: string
          example: "uuid-device-id"
        deviceType:
          type: string
          enum: [mobile, desktop, tablet]
          example: "mobile"
        appVersion:
          type: string
          example: "1.0.0"
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            accessToken:
              type: string
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            refreshToken:
              type: string
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            tokenType:
              type: string
              example: "Bearer"
            expiresIn:
              type: integer
              example: 3600
            user:
              $ref: '#/components/schemas/UserProfile'
        timestamp:
          type: string
          format: date-time
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "somchai@example.com"
        firstName:
          type: string
          example: "สมชาย"
        lastName:
          type: string
          example: "ใจดี"
        role:
          type: string
          enum: [CUSTOMER, AGENT, ADMIN]
          example: "CUSTOMER"
        permissions:
          type: array
          items:
            type: string
          example: ["VIEW_POLICIES", "CREATE_CLAIMS"]
    RegisterRequest:
      type: object
      required:
        - personalInfo
        - contactInfo
        - credentials
        - consent
      properties:
        personalInfo:
          $ref: '#/components/schemas/PersonalInfo'
        contactInfo:
          $ref: '#/components/schemas/ContactInfo'
        credentials:
          $ref: '#/components/schemas/Credentials'
        consent:
          $ref: '#/components/schemas/ConsentInfo'
    PersonalInfo:
      type: object
      required:
        - firstName
        - lastName
        - nationalId
        - dateOfBirth
        - gender
      properties:
        firstName:
          type: string
          example: "สมชาย"
        lastName:
          type: string
          example: "ใจดี"
        firstNameEn:
          type: string
          example: "Somchai"
        lastNameEn:
          type: string
          example: "Jaidee"
        nationalId:
          type: string
          pattern: '^[0-9]{13}$'
          example: "1234567890123"
        dateOfBirth:
          type: string
          format: date
          example: "1990-01-15"
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
          example: "MALE"
    ContactInfo:
      type: object
      required:
        - email
        - phoneNumber
        - address
      properties:
        email:
          type: string
          format: email
          example: "somchai@example.com"
        phoneNumber:
          type: string
          pattern: '^\+66[0-9]{9}$'
          example: "+66812345678"
        preferredLanguage:
          type: string
          enum: [th, en]
          default: th
          example: "th"
        address:
          $ref: '#/components/schemas/Address'
    Address:
      type: object
      required:
        - addressLine1
        - tambon
        - amphoe
        - province
        - postalCode
        - country
      properties:
        addressLine1:
          type: string
          example: "123/45 หมู่ 2"
        addressLine2:
          type: string
          example: "ถนนสุขุมวิท"
        tambon:
          type: string
          example: "คลองตัน"
        amphoe:
          type: string
          example: "วัฒนา"
        province:
          type: string
          example: "กรุงเทพมหานคร"
        postalCode:
          type: string
          pattern: '^[0-9]{5}$'
          example: "10110"
        country:
          type: string
          enum: [TH]
          default: TH
          example: "TH"
    Credentials:
      type: object
      required:
        - email
        - password
        - confirmPassword
      properties:
        email:
          type: string
          format: email
          example: "somchai@example.com"
        password:
          type: string
          minLength: 8
          pattern: '^(?=.*[A-Za-z])(?=.*\d)(?=.*[@$!%*#?&])[A-Za-z\d@$!%*#?&]'
          example: "SecurePass123!"
        confirmPassword:
          type: string
          example: "SecurePass123!"
    ConsentInfo:
      type: object
      required:
        - termsAndConditions
        - privacyPolicy
        - dataProcessingConsent
      properties:
        termsAndConditions:
          type: boolean
          example: true
        privacyPolicy:
          type: boolean
          example: true
        marketingConsent:
          type: boolean
          default: false
          example: false
        dataProcessingConsent:
          type: boolean
          example: true
    RegisterResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            customerId:
              type: string
              format: uuid
              example: "550e8400-e29b-41d4-a716-446655440000"
            verificationRequired:
              type: boolean
              example: true
            verificationMethods:
              type: array
              items:
                type: string
                enum: [EMAIL, SMS]
              example: ["EMAIL", "SMS"]
            message:
              type: string
              example: "Registration successful. Please verify your email and phone number."
        timestamp:
          type: string
          format: date-time
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    RefreshTokenResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            accessToken:
              type: string
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            expiresIn:
              type: integer
              example: 3600
        timestamp:
          type: string
          format: date-time
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: "somchai@example.com"
    ResetPasswordRequest:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
          example: "reset-token-uuid"
        newPassword:
          type: string
          minLength: 8
          example: "NewSecurePass123!"

    # Customer Management Schemas
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Request validation failed"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  code:
                    type: string
                  message:
                    type: string
                  messageEn:
                    type: string
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              format: uuid
            path:
              type: string
...